importccl: move csv reader to own file
Release note: none.
